- email validation server side - Minors
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/branches/com_virtuemart.3.0.12.4@9554 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/models/updatesmigration.php
@@ -254,7 +254,7 @@ class VirtueMartModelUpdatesMigration extends VmModel {
 	 * @param String $dst path
 	 * @param String $type modules, plugins, languageBE, languageFE
 	 */
-	static public function recurse_copy($src,$dst ) {
+	static public function recurse_copy($src,$dst,$delete = true ) {
 
 		if(!class_exists('JFolder'))
 			require(VMPATH_LIBS.DS.'joomla'.DS.'filesystem'.DS.'folder.php');
@@ -274,7 +274,7 @@ class VirtueMartModelUpdatesMigration extends VmModel {
 							self::recurse_copy($src .DS. $file,$dst .DS. $file);
 						}
 						else {
-							if(JFile::exists($dst .DS. $file)){
+							if($delete and JFile::exists($dst .DS. $file)){
 								if(!JFile::delete($dst .DS. $file)){
 									$app = JFactory::getApplication();
 									$app -> enqueueMessage('Couldnt delete '.$dst .DS. $file);
